Update PEEV to handle updated type forwarding
The key change is to visit lambda parameters before the body, and
then to iteratively visit the body with a parameter rebinding
visitor.

This commit also reduces the surface area and responsibility of
the `TypeCache`. It should only be concerned with maintaining
visitor-local mappings for types and parameters.

`WithExpression` was also refactored to improve the debugging
experience. Currently, if created via the `.With(...)` extension
method, the expression's `.Reduce()` is called before passing back
the queryable.